==> notebooks/4_team_csv.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import plotly.express as px

# Creating the Season CSV Information

Using the CSV from scraping stats, and filtering it to display info based on the statistic the user chooses

season_stats = pd.read_csv('data/processed/season_stats.csv')

season_stats[['FG Made', 'FG Attempted']] = season_stats['FG: Made-Attempted'].str.split('-', n=1, expand=True)
season_stats[['3PT Made', '3PT Attempted']] = season_stats['3PT: Made-Attempted'].str.split('-', n=1, expand=True)
season_stats[['FT Made', 'FT Attempted']] = season_stats['FT: Made-Attempted'].str.split('-', n=1, expand=True)

season_stats
# Remove Fg: Made-Attempted when defining columns


print(season_stats.columns)

season_stats_uiuc = season_stats[season_stats['Team'] == 'Illinois']



fig = px.line(season_stats_uiuc, x=season_stats_uiuc['Season'], y='Points Per Game', title='Points Per Game by Season')
fig
